requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
colorama==0.4.6
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
